+++ NEWS.md
@@ -7,6 +7,9 @@
   - don't crash when first unrecognized argument is also a handler
     name (boolean, file, finished, flag, etc...)
 
+  - don't hang when help text option table formats long option name
+    on its own line before indented description.
+
 
 ## Noteworthy changes in release 1.1.1 (2016-02-07) [stable]